refactor: extract KgsWidgetDataSource
